refining tests, and removing redundancy in discovery of research tools
