+++ tests/TEST_INPUTS/external_repo_template/plots/external_plot.py
@@ -1,0 +1,6 @@
+"""Dummy external plot for registry tests."""
+
+
+class ExternalPlot:
+    """Dummy external plot for registry tests only."""
+    pass
