+++ tests/TEST_INPUTS/external_repo_template/projects/external_project.py
@@ -1,0 +1,6 @@
+"""Dummy external project for registry tests."""
+
+
+class ExternalProject:
+    """Dummy external project for registry tests only."""
+    pass
