+++ tests/TEST_INPUTS/external_repo_template/reports/external_report.py
@@ -1,0 +1,6 @@
+"""Dummy external report for registry tests."""
+
+
+class ExternalReport:
+    """Dummy external report for registry tests only."""
+    pass
